Refactor main package to pkg/rootcommand
- enables testing of main package functions
- slims main package
- allows reuse of main package functions in docs main package which generates docs
